requirements.txt
# Web API
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
httpx==0.27.0

# Database
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1

# Elasticsearch
elasticsearch[async]==8.12.0

# Redis / Celery
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# Authentication
authlib==1.3.0
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Azure
azure-storage-blob==12.19.1
azure-identity==1.15.0

# Document Processing
extract-msg==0.48.5
python-pptx==0.6.23
openpyxl==3.1.2
python-docx==1.1.0
pypdf==4.0.2
pdfplumber==0.10.4
chardet==5.2.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1
hashlib-additional==1.0.0
